my custom keymap for my pinky3
It takes a lot of inspiration from the default keymap that came with my
planck EZ in terms of where numbers and symbols are are, which layer
maps to shifted things or not, etc. I also stole an idea from my redox
wireless where I used layertap for things like backspace and whatnot,
which is really growing on me, so there are even a few unused keys.