cmd-{ and cmd-} configs
these were really awkward with the normal keymap from the planck, this
is a little better. It doesn't make a lot of sense though if you think
about it as a shifted combo of `cmd-[` and `cmd-]`, but maybe I'll
figure out an alternative later. I thought about putting those on the jk
keys on the same layer and might still do. we'll see