bongocat wpm meter
works like a champ. Though I'd like to see if I can get my led
animations back and get rid of the wpm and just have a keystroke
counter. I'll look at the wpm code to see how it counts things and
simplify that to just a straight up counter with a timeout (probably
when bongocat goes back to rest state, or when the OLEDs got to rest
state) that saves the counter to eeprom so it doesn't get lost between reboots.